SingleplayerMode
Seeing the current requirements for the latest build (having two screens to play) I decided to see about a "singleplayer" mode that was also different from the previous split-screen mode we had at the start.

Obviously, the game might not go in this direction / there might not be a real need for a single-player mode but I figured I'd implement it as more of a proof of concept. 

This mode uses PIP or Picture In Picture to display the secondary cameras over the player's HUD. This means the player must choose between being able to see what's actually in front of them or what the cameras can see.